only continue if annotation list does not already exist

+++ lib/tasks/hocr_annotate.rake
@@ -14,11 +14,11 @@ namespace :ocracoke do
       # This only works in the case when images have an underscore and resources don't
       basename = File.basename directory
       hocr = File.join directory, basename + '.hocr'
-      if File.exist? hocr
+      outfile = File.join directory, basename + '-annotation-list.json'
+      if File.exist?(hocr) && !File.exist?(outfile)
         puts hocr
         hr = HocrOpenAnnotationCreator.new File.expand_path(hocr)
         annotations = hr.annotation_list
-        outfile = File.join directory, basename + '-annotation-list.json'
         File.open(outfile, 'w') do |fh|
           fh.puts annotations.to_json
         end
